refactor: 패키지 구조 수정
- data, domain 패키지 ui 폴더 밖으로 이동
