+++ app/src/main/java/com/example/team25/domain/model/AccompanyInfo.kt
@@ -1,4 +1,4 @@
-package com.example.team25.ui.domain.model
+package com.example.team25.domain.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
