+++ app/src/main/java/com/example/team25/domain/model/Gender.kt
@@ -1,0 +1,5 @@
+package com.example.team25.domain.model
+
+enum class Gender {
+    MALE, FEMALE
+}
