--- app/src/main/java/com/example/team25/ui/domain/model/Gender.kt
@@ -1,5 +1,0 @@
-package com.example.team25.ui.domain.model
-
-enum class Gender {
-    MALE, FEMALE
-}
